requirements.txt
streamlit==1.45.0
openai==1.77.0
pandas==2.2.3
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
torch==2.7.0
Pillow==11.2.1
ultralytics==8.3.123
